action/ obs space

+++ tria-rl/tria_rl/envs/tria_climate.py
@@ -43,12 +43,12 @@ class TriaClimateEnv(gym.Env):
         low = np.array([self.metadata['t_min'], self.metadata['h_min'], self.metadata['a_min']])#.astype(np.int32)
         high = np.array([self.metadata['t_max'], self.metadata['h_max'], self.metadata['a_max']])#.astype(np.int32)
 
-        self.observation_space = spaces.Box(low, high, shape=(3,))
+        self.observation_space = spaces.Box(low, high, shape=(3,), dtype=np.int32)
 
         # We have 2 actions, corresponding to "on", "off"
         # spaces.MultiDiscrete(np.array([self.metadata['action_states'], self.metadata['action_states'], self.metadata['action_states'], self.metadata['action_states'], self.metadata['action_states']]))
         #self.action_space = tuple((spaces.Discrete(2),spaces.Discrete(2),spaces.Discrete(2),spaces.Discrete(2),spaces.Discrete(2)))
-        self.action_space = spaces.Box(low=0, high=1, shape=(5,))
+        self.action_space = spaces.Box(low=0, high=1, shape=(5,), dtype=np.int32)
 
 
         self.mean = [self.metadata['range_dict'][0][0] + self.metadata['range_dict'][0][1] // 2,
